Image_Classification
The idea is that identifying different species is a more challenging problem than identifying different animals or even animals from objects. The dataset contains 225 bird species with training, validation and test images. 
The dataset was downloaded from Kaggle (https://www.kaggle.com/gpiosenka/100-bird-species).
Thus the goal is to find a systematic way to choose between a few different NN models and parameters and use the best model to classify different bird species and apply the model to an image completion problem in order to compare it to other ML classifiers. The best model is selected by validation accuracy/loss as well as training time.
Classes to add layers to NNs are used to define a model builder that can create different models:
A customized grid search will also be performed along the parameters of filter size, kernel size and learning rate.
Because this can take very long when many parameters are used, an algorithm has to be built that compares different configurations and models.  After e.g. 5, 10 and 15 epochs models and parameters that don’t meet a certain accuracy threshold have to be filtered out. Models can be sorted by accuracy or loss to find the best performing one. 
The best performing model will then be used to classify different bird species. 
The classification result can be shown with associated probabilities for the same working base model 
The model can then be refined by performing a feature selection. In the lecture it was mentioned that non-negative matrix factorization (NMF) is performing better than single-value decomposition(SVD) because of the lack of interpretability for negative feature values in images (https://mlexplained.com/2017/12/28/a-practical-introduction-to-nmf-nonnegative-matrix-factorization/).
Challenges are that both NMF and SVD yield reconstructed arrays that are not rectangular anymore (224x224) and it will have to be examined how compatible this is with Conv2d in Keras. Instead, principal component analysis (PCA) was performed. For this, the RGB image had to decomposed in a red, a green and a blue channel and PCA was performed for each channel before combining it again to reconstruct the image
It is interesting to investigate if the NN takes less training time by transforming all images with PCA while maintaining good accuracy.
Image completion was already done with KNN on a few sample images, for which the classifier was trained to predict the lower half of an image using the upper half.
The NN model could be used to for image completion with RGB images and be compared (e.g. analytically element-wise or by visual inspection) to other classifiers. For this application NMF will also be performed on the simpler color scheme to make the method more efficient.